requirements.txt
torch==2.5.1+cu118
torchvision==0.20.1+cu118  
torchaudio==2.5.1+cu118
pytorch-lightning==2.0.0
transformers==4.27.1
tokenizers==0.13.3
faiss-cpu==1.11.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
ollama==0.5.3
psutil==7.0.0
tqdm==4.67.1
numpy==2.1.2
pydantic==2.11.7
requests==2.32.4